fix: decode Brady gap counts as INT4 in constitution screens
gaps_found is an INT column; sqlx rejected i64 (INT8) during
POST /constitution/screen on the seeded demo case.

+++ crates/vi-constitution/src/db.rs
@@ -338,7 +338,7 @@ pub async fn screen_case(
     .fetch_one(pool)
     .await?;
 
-    let brady_gaps: i64 = sqlx::query_scalar(
+    let brady_gaps: i32 = sqlx::query_scalar(
         "SELECT COALESCE((SELECT gaps_found FROM brady_recon_runs
           WHERE case_id = $1 ORDER BY run_at DESC LIMIT 1), 0)",
     )
